Refactor recent documents listing. - Moved to dedicated controller - Added more unit tests
