+++ tests/DocumentsTest.php
@@ -84,34 +84,6 @@ class DocumentsTest extends BrowserKitTestCase
         ];
     }
 
-    public function recent_date_range_provider()
-    {
-        return [
-            ['today'],
-            ['yesterday'],
-            ['currentweek'],
-            ['currentmonth'],
-        ];
-    }
-
-    public function recent_items_per_page_provider()
-    {
-        return [
-            [5],
-            [10],
-            [15],
-            [25],
-            [50],
-        ];
-    }
-
-    public function recent_sorting_provider()
-    {
-        return [
-            ['a', 'ASC'],
-            ['d', 'DESC'],
-        ];
-    }
 
     /**
      * Test that the route for documents.show is not leaking private documents to anyone
@@ -923,187 +895,6 @@ class DocumentsTest extends BrowserKitTestCase
         $this->assertViewHas('clear_filter_url');
 
         $this->assertEquals($url, $view->clear_filter_url);
-    }
-
-    /**
-     * @dataProvider recent_date_range_provider
-     */
-    public function testRecentPageRangeSelection($range)
-    {
-        Flags::enable(Flags::UNIFIED_SEARCH);
-
-        $this->withKlinkAdapterFake();
-
-        $user = $this->createUser(Capability::$PROJECT_MANAGER_NO_CLEAN_TRASH);
-        $this->actingAs($user);
-        
-        // goto link, see linked page
-
-        $url = route('documents.recent', ['range' => $range]);
-
-        $this->visit($url)->seePageIs($url);
-
-        $this->assertViewHas('range', $range);
-
-        $user = $user->fresh();
-        $this->assertEquals($range, $user->optionRecentRange());
-    }
-    
-    /**
-     * Test Items per page option is honored
-     *
-     * @dataProvider recent_items_per_page_provider
-     */
-    public function testRecentPageItemsPerPageSelection($items_per_page)
-    {
-        Flags::enable(Flags::UNIFIED_SEARCH);
-
-        $this->withKlinkAdapterFake();
-
-        $user = $this->createUser(Capability::$PROJECT_MANAGER_NO_CLEAN_TRASH);
-        $this->actingAs($user);
-
-        for ($i=0; $i < $items_per_page; $i++) {
-            $this->createDocument($user);
-        }
-
-        $url = route('documents.recent').'?n='.$items_per_page ;
-
-        $this->visit($url)->seePageIs($url);
-        $user = $user->fresh();
-        $this->assertEquals($items_per_page, $user->optionItemsPerPage());
-        
-        $this->assertEquals($items_per_page,
-            $this->response->original->documents->values()->collapse()->count());
-    }
-    
-    /**
-     * Test sorting option is honored
-     *
-     * @dataProvider recent_sorting_provider
-     */
-    public function testRecentPageSortingSelection($option, $expected_value)
-    {
-        $this->withKlinkAdapterFake();
-
-        $user = $this->createUser(Capability::$PROJECT_MANAGER_NO_CLEAN_TRASH);
-        $this->actingAs($user);
-
-        // for ($i=0; $i < $items_per_page; $i++) {
-        //     $this->createDocument($user);
-        // }
-
-        $url = route('documents.recent').'?o='.$option ;
-
-        $this->visit($url)->seePageIs($url);
-        
-        $this->assertViewHas('order', $expected_value);
-    }
-    
-    /**
-     * Test recent page contains expected documents
-     *
-     * - last updated by user
-     * - last shared with me
-     */
-    public function testRecentPageContainsExpectedDocuments($range = 'today')
-    {
-        $this->withKlinkAdapterFake();
-
-        $user = $this->createUser(Capability::$PROJECT_MANAGER_NO_CLEAN_TRASH);
-        $user2 = $this->createUser(Capability::$PROJECT_MANAGER_NO_CLEAN_TRASH);
-        
-        $this->actingAs($user);
-
-        $documents = collect(); // documents created by $user
-        $documents_user2 = collect();  // documents created by $user2
-
-        $count_documents_by_me = 5;
-        $count_documents_shared_with_me = 1;
-        $count_documents_in_project = 1;
-
-        // create some documents for $user
-
-        for ($i=0; $i < $count_documents_by_me; $i++) {
-            $documents->push($this->createDocument($user));
-        }
-        
-
-        $doc = null;
-        
-        // create a project using $user2, add 1 document in the project
-        $project1 = $this->createProject(['user_id' => $user2->id]);
-        $project1->users()->attach($user->id);
-        $doc = $this->createDocument($user2);
-        $service = app('Klink\DmsDocuments\DocumentsService');
-        $service->addDocumentToGroup($user2, $doc, $project1->collection);
-        $doc = $doc->fresh();
-        $documents_user2->push($doc);
-
-        // create a second user, share with the first one a couple of documents
-        for ($i=0; $i < $count_documents_shared_with_me; $i++) {
-            $doc = $this->createDocument($user2);
-
-            $doc->shares()->create([
-                        'user_id' => $user2->id,
-                        'sharedwith_id' => $user->id, //the id
-                        'sharedwith_type' => get_class($user), //the class
-                        'token' => hash('sha512', $doc->id),
-                    ]);
-            $documents_user2->push($doc);
-        }
-
-        // grab the last from $documents and change its updated_at to yesterday
-        // (wrt the selected $range)
-        $last = $documents->last();
-
-        $last->updated_at = Carbon::yesterday();
-
-        $last->timestamps = false; //temporarly disable the automatic upgrade of the updated_at field
-
-        $last->save();
-
-        $url = route('documents.recent', ['range' => $range]);
-
-        $this->visit($url)->seePageIs($url);
-
-        $this->assertEquals(($count_documents_by_me - 1) + $count_documents_shared_with_me + $count_documents_in_project,
-            $this->response->original->documents->values()->collapse()->count());
-    }
-
-    public function testRecentPageSearchWithFilters()
-    {
-        Flags::enable(Flags::UNIFIED_SEARCH);
-
-        $docs = factory('KBox\DocumentDescriptor', 10)->create();
-
-        $adapter = $this->withKlinkAdapterFake();
-
-        // prepare the request
-        $searchRequest = KlinkSearchRequest::build('*', 'private', 1, 1, [], []);
-        
-        // prepare some fake results
-        $adapter->setSearchResults('private', KlinkSearchResults::fake($searchRequest, []));
-
-        $user = $this->createUser(Capability::$PROJECT_MANAGER_NO_CLEAN_TRASH);
-        $this->actingAs($user);
-
-        $url = route('documents.recent').'?s=hello';
-
-        $this->visit($url)->seePageIs($url);
-        
-        $this->assertViewHas('search_replica_parameters', ['s' => 'hello']);
-
-        $this->see(route('documents.recent', ['range' => 'currentweek', 'n' => 12, 's' => 'hello']));
-        $this->see(route('documents.recent', ['range' => 'currentweek', 'n' => 24, 's' => 'hello']));
-        $this->see(route('documents.recent', ['range' => 'currentweek', 'n' => 50, 's' => 'hello']));
-        
-        $this->see(route('documents.recent', ['range' => 'today', 's' => 'hello']));
-        $this->see(route('documents.recent', ['range' => 'yesterday', 's' => 'hello']));
-        $this->see(route('documents.recent', ['range' => 'currentweek', 's' => 'hello']));
-        $this->see(route('documents.recent', ['range' => 'currentmonth', 's' => 'hello']));
-
-        $this->see('search-form');
     }
 
     /**
